+++ CHANGELOG.md
@@ -1,3 +1,7 @@
+## Next Version
+### Bug fixes
+* Fix extra out parameters in the fragment shader not getting added to the function declaration. [(more)](https://github.com/d4rkc0d3r/d4rkAvatarOptimizer/issues/10)
+
 ## v1.5
 ### Bug fixes
 * Fix crash when an EditorOnly tagged game object was in the exclusions list.